Refactor invoices to their own module
