+++ saleor/graphql/core/utils/error_codes.py
@@ -6,8 +6,9 @@ from ....checkout.error_codes import CheckoutErrorCode
 from ....core.error_codes import ShopErrorCode
 from ....discount.error_codes import DiscountErrorCode
 from ....giftcard.error_codes import GiftCardErrorCode
+from ....invoice.error_codes import InvoiceErrorCode
 from ....menu.error_codes import MenuErrorCode
-from ....order.error_codes import InvoiceErrorCode, OrderErrorCode
+from ....order.error_codes import OrderErrorCode
 from ....payment.error_codes import PaymentErrorCode
 from ....plugins.error_codes import PluginErrorCode
 from ....product.error_codes import ProductErrorCode
