+++ saleor/graphql/invoice/enums.py
@@ -1,0 +1,5 @@
+class InvoiceStatus:
+    PENDING = "pending"
+    PENDING_DELETE = "pending_delete"
+    DELETED = "deleted"
+    READY = "ready"
