+++ saleor/order/__init__.py
@@ -103,19 +103,3 @@ class OrderEventsEmails:
         (FULFILLMENT, "The fulfillment confirmation email was sent"),
         (DIGITAL_LINKS, "The email containing the digital links was sent"),
     ]
-
-
-class InvoiceEvents:
-    REQUESTED = "requested"
-    REQUESTED_DELETION = "requested_deletion"
-    CREATED = "created"
-    DELETED = "deleted"
-    SENT = "sent"
-
-    CHOICES = [
-        (REQUESTED, "The invoice was requested"),
-        (REQUESTED_DELETION, "The invoice was requested for deletion"),
-        (CREATED, "The invoice was created"),
-        (DELETED, "The invoice was deleted"),
-        (SENT, "The invoice has been sent"),
-    ]
