+++ saleor/order/error_codes.py
@@ -24,11 +24,3 @@ class OrderErrorCode(Enum):
     ZERO_QUANTITY = "zero_quantity"
     INSUFFICIENT_STOCK = "insufficient_stock"
     DUPLICATED_INPUT_ITEM = "duplicated_input_item"
-
-
-class InvoiceErrorCode(Enum):
-    REQUIRED = "required"
-    NOT_READY = "not_ready"
-    URL_OR_NUMBER_NOT_SET = "url_or_number_not_set"
-    NOT_FOUND = "not_found"
-    INVALID_STATUS = "invalid_status"
